+++ integration-tests/global-setup/setup.js
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+const crypto = require('crypto');
+const fs = require('fs');
+const mkdirp = require('mkdirp');
+const os = require('os');
+const path = require('path');
+
+const DIR = path.join(os.tmpdir(), 'jest-global-setup');
+
+module.exports = function() {
+  return new Promise((resolve, reject) => {
+    mkdirp.sync(DIR);
+    const fileId = crypto.randomBytes(20).toString('hex');
+    fs.writeFileSync(path.join(DIR, fileId), 'setup');
+    resolve();
+  });
+};